+++ compiler/builtins.rs
@@ -1,5 +1,8 @@
 use crate::{Object, Objects};
-use sflynlang_parser::{Error, Position};
+use sflynlang_parser::{
+    ast::{DataType, DataTypes},
+    Error, Position,
+};
 
 pub fn print(
     objects: Vec<Object>,
@@ -48,4 +51,34 @@ pub fn debug(
     }
 
     Ok(Object::new(object.get_position(), Objects::Void))
+}
+
+pub fn check_builtin(
+    key: String,
+    position: Position,
+    arguments: Vec<DataType>,
+) -> Result<DataType, Error> {
+    match key.as_str() {
+        "print" | "debug" => {
+            if arguments.len() != 1 {
+                return Err(Error::new_expect_arguments(
+                    position,
+                    1,
+                    arguments.len(),
+                ));
+            }
+
+            if arguments[0].node == DataTypes::String {
+                Ok(DataType::new(position, DataTypes::Void))
+            } else {
+                Err(Error::new_expect_type(
+                    position,
+                    "string",
+                    &arguments[0].node.to_string(),
+                ))
+            }
+        }
+
+        _ => Err(Error::new_unknown_identifier(position, key)),
+    }
 }