+++ compiler/lib.rs
@@ -1,7 +1,8 @@
-mod builtins;
+pub mod builtins;
 mod environment;
 mod evaluator;
 mod objects;
+mod typechecker;
 
 pub use environment::{Environment, Store};
 pub use objects::{Object, Objects};
@@ -15,6 +16,10 @@ pub fn run(statements: Vec<Statement>, debug_mode: bool, file: &File) -> i32 {
 
     for statement in statements.iter() {
         if let Err(error) =
+            typechecker::check_statement(statement, &mut environment)
+        {
+            environment.add_error(error);
+        } else if let Err(error) =
             evaluator::evaluate_statement(statement, &mut environment)
         {
             environment.add_error(error);